+++ eos_tax/__init__.py
@@ -1,5 +1,5 @@
 default_app_config = "eos_tax.apps.EosTaxConfig"
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"
 VERSION = __version__